Move get balance (secp256k1) related prefab converters into a single file
